Move tests from test_deserialize_siteroot.py to test_site_deserializer.py.
